scripts/requirements.txt
streamlit==1.28.1
PyPDF2==3.0.1
python-docx==0.8.11
pandas==2.1.1
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0
